Fix SpellBad highlight group of dim color scheme using vim-aftercolors
My current vim colorscheme, dim, specifies:

  :highlight SpellBad
  SpellBad       xxx ctermbg=2 gui=undercurl guisp=Red

When setting terminal colors to the solarized theme (in alacritty), dim
specifies `ctermbg=2`, which makes the highlighted misspelling
illegible.

This seems odd to use a plugin to fix, but this can't trivially be fixed
by modifying the vimrc without adding a fair amount of complexity.
Usually, you'd want to ensure color scheme overrides are only applied to
the intended scheme, and you need to make sure your fix is run after
each time the color scheme is loaded, which is roughly every time a file
is loaded.

The other solution I considered is forking the vim-dim colorscheme
rather than depending on it as an external plugin. Should the current
solution give me any grief, this is what I'd do next (and may do
anyways).